# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API HUB-UMKM Documentation",
        "description": "Documentation for HUB PHB UMKM",
        "contact": {
            "email": "jamal.apriadi@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Laravel Swagger API server"
        }
    ],
    "paths": {
...
        "/api/product/sync": {
            "post": {
                "tags": [
                    "External Integration"
                ],
                "summary": "Sync a single product from external store (by seller client ID and client secret)",
                "description": "Digunakan oleh toko eksternal untuk mengirim data produk ke marketplace pusat.",
                "operationId": "ec3acc516fa27356e96e34a4d123fb3d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "client_id",
                                    "seller_product_id",
                                    "name",
                                    "price",
                                    "stock"
                                ],
                                "properties": {
                                    "client_id": {
                                        "type": "string",
                                        "example": "abc123"
                                    },
                                    "seller_product_id": {
                                        "type": "string",
                                        "example": "SP001"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Tas Ransel Hitam"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Tas untuk kerja & sekolah"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 159000
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "sku": {
                                        "type": "string",
                                        "example": "SKU-001"
                                    },
                                    "image_url": {
                                        "type": "string",
                                        "example": "https://example.com/image.jpg"
                                    },
                                    "weight": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 0.8
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product synced successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product synced successfully."
                                        },
                                        "product_id": {
                                            "type": "integer",
                                            "example": 12
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Seller not found"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}